% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafplot.R
\name{leafplot}
\alias{leafplot}
\title{Land and Water Plot}
\usage{
leafplot(u, pw, port, usePlotly = FALSE)
}
\arguments{
\item{schema}{name of the database schema to use. Default: "soil_survey_data"}

\item{max}{maximum number of records to limit the resultst to. Default: 20, max: 100}
}
\description{
Plots farm and soil related stats
}
\examples{
\dontrun{
leafplot(u = "db_user", pw = "db_password", port = db_port)
leafplot(u = "db_user", pw = "db_password", port = db_port, usePlotly = TRUE)
}
}
\author{
Ashish Singh https://github.com/git-ashish
}
