% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwmplotly.R
\name{lwmplotly}
\alias{lwmplotly}
\title{Land and Water Plot using Plotly}
\usage{
lwmplotly(schema = "soil_survey_data", max = 20, u, pw, port)
}
\arguments{
\item{schema}{name of the database schema to use. Default: "soil_survey_data"}

\item{max}{maximum number of records to limit the resultst to. Default: 20, max: 100}
}
\description{
Plots farm and soil related stats
}
\examples{
\dontrun{
lwmplot(max = 10)
lwmplot(max = 40)
lwmplot(schema = "farms", max = 30)
}
}
\author{
Ashish Singh https://github.com/git-ashish
}
